Ai_Intern/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# PDF Processing
pypdf==3.17.4

# Text Processing
nltk==3.8.1
tiktoken==0.5.2

# Embeddings
sentence-transformers==2.2.2
torch>=2.2.0

# Vector Database
chromadb==0.4.18

# LLM Integration
openai==1.3.5

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Jupyter Notebook
jupyter==1.0.0
ipykernel==6.26.0
notebook==7.0.6

# Utilities
requests==2.31.0

